Redesigns the layout and style to create a new div that contains a nested stack.
When the escher stack is displayed, it creates a new div that is layered on top of the base layer. Each subsequent view is nested in the div. As the stack grows, each previous layer's height is adjusted to keep things tidy.

A resize event handler is hooked to the window resize event, to resize the escher stack. It's only registered if there are multiple layers in the stack.